Move tests from `components/unread-articles-indicator/test` to `test/unread-articles-indicator`
 🐿 v2.10.2
